feat(logs) : implémentation du module de journalisation avec CreateLogController et le service LogRecorder
- Ajout de CreateLogController pour gérer la création de journaux via le point de terminaison POST /api/logs.
- Introduction du service LogRecorder pour le traitement et l'enregistrement des entrées de journal.
- Création des entités LogEntry, LogEntryTag, LogEnv, LogLevel, LogRoute, LogSource et LogTag avec les relations appropriées.
- Implémentation de référentiels pour chaque entité afin de faciliter les interactions avec la base de données.
- Ajout de règles de réécriture Apache pour le contrôleur frontal Symfony dans le fichier .htaccess.
- Mise à jour de la documentation relative aux journaux afin de refléter les changements apportés à la structure et aux fonctionnalités du service.
