+++ api/postman/corbidev-api-logs.postman_collection.json
@@ -48,7 +48,7 @@
   "variable": [
     {
       "key": "baseUrl",
-      "value": "http://127.0.0.1:8000"
+      "value": "http://api.corbisier.test"
     },
     {
       "key": "apiKey",
