+++ api/public/.htaccess
@@ -1,0 +1,17 @@
+# Apache rewrite rules for Symfony front controller
+
+<IfModule mod_negotiation.c>
+    Options -MultiViews
+</IfModule>
+
+<IfModule mod_rewrite.c>
+    RewriteEngine On
+
+    # Let Apache serve existing files directly
+    RewriteCond %{REQUEST_FILENAME} -f
+    RewriteRule ^ - [L]
+
+    # Send all other requests to Symfony
+    RewriteRule ^ index.php [QSA,L]
+
+</IfModule>
