+++ api/src/RessLogs/Entity/LogEntry.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Entity\Log;
+namespace App\RessLogs\Entity;
 
-use App\Repository\Log\LogEntryRepository;
+use App\RessLogs\Repository\LogEntryRepository;
 use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
